+++ ExpenseTracking.Api/Impl/Cqrs/PaymentMethod.cs
@@ -5,7 +5,7 @@ using MediatR;
 namespace ExpenseTracking.Api.Impl.Cqrs.PaymentMethod;
 
 public record GetAllPaymentMethodsQuery : IRequest<ApiResponse<List<PaymentMethodResponse>>>;
-public record GetPaymentMethodByIdQuery(int Id) : IRequest<ApiResponse<PaymentMethodResponse>>;
+public record GetPaymentMethodByIdQuery(long Id) : IRequest<ApiResponse<PaymentMethodResponse>>;
 public record CreatePaymentMethodCommand(PaymentMethodRequest PaymentMethod) : IRequest<ApiResponse<PaymentMethodResponse>>;
-public record UpdatePaymentMethodCommand(int Id, UpdatePaymentMethodRequest PaymentMethod) : IRequest<ApiResponse>;
-public record DeletePaymentMethodCommand(int Id) : IRequest<ApiResponse>;
+public record UpdatePaymentMethodCommand(long Id, UpdatePaymentMethodRequest PaymentMethod) : IRequest<ApiResponse>;
+public record DeletePaymentMethodCommand(long Id) : IRequest<ApiResponse>;
